+++ docs/fr_FR/changelog.md
@@ -4,6 +4,10 @@
 >
 >S'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+# 04/09/2024
+- Mise à jour de la librairie en version 1.0.5
+- Nouveau champs calculés : Puissance active totale et par phases, Intensité totale en triphasé.
+
 # 04/12/2023
 - Correction erreur de non suppression de l'ancien topic lors d'un changement de ce dernier.